requirements.txt
gspread
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
requests
yt-dlp
undetected-chromedriver
selenium
openai
pandas
numpy
opencv-python
Pillow
librosa
soundfile
pyrubberband
